feat: 상품 주문서 생성 기능 구현

+++ src/main/java/kr/hhplus/be/server/domain/coupon/Coupon.java
@@ -10,5 +10,5 @@ public class Coupon {
     @GeneratedValue
     private Long id;
     private String name;
-    private Integer discountPercent;
+    private Long discountPercent;
 }
